NEW_Panns/requirements.txt
# PANNs Heart Sound Classifier - Requirements

# Deep Learning Framework
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
scipy>=1.10.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning Utils
scikit-learn>=1.3.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
